w3c-reports/12092014/efl/runned-tests/websockets/cookies/005-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/set-cookie_http?1410542078271.0.9440466498490423' failed: Compressed bit must be 0 if no negotiated deflate-frame extension


FAIL WebSockets: setting HttpOnly cookies in ws response, checking ws request assert_unreached: Reached unreachable code(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_unreached@http://localhost:8800/resources/testharness.js:692:16
http://localhost:8800/websockets/cookies/005.html:24:56
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
